flaskapi/requirements.txt
flask
pymongo
python-dotenv
bcrypt
pyjwt
gunicorn